requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cycler==0.11.0
dataclasses==0.6
fonttools==4.38.0
future==0.18.3
idna==3.4
imageio==2.26.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wget==3.2
